% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gasoline}
\alias{gasoline}
\title{US finished motor gasoline product supplied.}
\format{
Time series object of class \code{ts}.
}
\source{
US Energy Information Administration.
}
\description{
Weekly data beginning 2 February 1991, ending 20 January 2017. Units are
"million barrels per day".
}
\examples{

autoplot(gasoline, xlab = "Year")

}
\keyword{datasets}
