% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{goog}
\alias{goog}
\alias{goog200}
\title{Daily closing stock prices of Google Inc}
\format{
Daily time series of class \code{ts}.
}
\source{
\url{https://finance.yahoo.com/quote/GOOG/history/}
}
\description{
Closing stock prices of GOOG from the NASDAQ exchange, for 1000 consecutive
trading days between 25 February 2013 and 13 February 2017. Adjusted for
splits. \code{goog200} contains the first 200 observations from \code{goog}.
}
\examples{

autoplot(goog)

}
\keyword{datasets}
