% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uschange}
\alias{uschange}
\alias{usconsumption}
\title{Growth rates of personal consumption and personal income in the USA.}
\format{
Time series object of class \code{ts}.
}
\source{
Federal Reserve Bank of St Louis.
}
\description{
Percentage changes in quarterly personal consumption expenditure, personal
disposable income, production, savings and the unemployment rate for the US,
1960 to 2016.
}
\examples{

autoplot(uschange, facet = TRUE)

}
\keyword{datasets}
